# schemas/research-surface-api-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "CycleStatus": {
        "description": "Terminal outcomes of one synchronous monitoring cycle.",
        "enum": [
          "NO_CHANGE",
          "ALERTS_EMITTED",
          "ATTENTION_REQUIRED",
          "FAILED"
        ],
        "title": "CycleStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MonitoringOperationsActivationV1": {
        "additionalProperties": false,
        "description": "The one projected activation: active when unfinished, else latest terminal.",
        "properties": {
          "activation_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Activation Id",
            "type": "string"
          },
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cycle_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Cycle Id",
            "type": "string"
          },
          "intent_content_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Intent Content Sha256",
            "type": "string"
          },
          "kind": {
            "enum": [
              "ACTIVE",
              "LATEST_TERMINAL"
            ],
            "title": "Kind",
            "type": "string"
          },
          "receipt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MonitoringOperationsReceiptV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Request Fingerprint",
            "type": "string"
          },
          "runner_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Runner Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "activation_id",
          "runner_id",
          "cycle_id",
          "as_of",
          "created_at",
          "request_fingerprint",
          "intent_content_sha256"
        ],
        "title": "MonitoringOperationsActivationV1",
        "type": "object"
      },
      "MonitoringOperationsCycleV1": {
        "additionalProperties": false,
        "description": "Terminal D1 cycle facts for the projected activation, when committed.",
        "properties": {
          "alert_batch_content_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Alert Batch Content Sha256",
            "type": "string"
          },
          "alert_batch_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Alert Batch Id",
            "type": "string"
          },
          "alert_count": {
            "minimum": 0.0,
            "title": "Alert Count",
            "type": "integer"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/MonitoringAlertV1"
            },
            "maxItems": 1024,
            "title": "Alerts",
            "type": "array"
          },
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "cycle_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Cycle Id",
            "type": "string"
          },
          "event_batch_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Batch Id"
          },
          "failure_code": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "monitoring_run_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monitoring Run Id"
          },
          "next_state_id": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next State Id"
          },
          "pointer_published": {
            "title": "Pointer Published",
            "type": "boolean"
          },
          "result_content_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Result Content Sha256",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CycleStatus"
          },
          "watchlist_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Watchlist Id",
            "type": "string"
          }
        },
        "required": [
          "cycle_id",
          "status",
          "as_of",
          "watchlist_id",
          "alert_batch_id",
          "alert_batch_content_sha256",
          "result_content_sha256",
          "pointer_published",
          "alert_count"
        ],
        "title": "MonitoringOperationsCycleV1",
        "type": "object"
      },
...
      "MonitoringOperationsDeliveryV1": {
        "additionalProperties": false,
        "description": "One D2B delivery bound to the projected activation.\n\n``state.attempt_count`` counts persisted attempt outcomes while\n``dispatch_claim_count`` counts consumed authorized outbound transport\nslots; after an orphaned dispatch the two legitimately differ and the\nunresolved slot numbers stay explicit (6-D2B-R2).",
        "properties": {
          "activation_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Activation Id",
            "type": "string"
          },
          "alert_batch_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Alert Batch Id",
            "type": "string"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/MonitoringOperationsDeliveryAttemptV1"
            },
            "maxItems": 64,
            "title": "Attempts",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cycle_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Cycle Id",
            "type": "string"
          },
          "delivery_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Delivery Id",
            "type": "string"
          },
          "destination_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Destination Id",
            "type": "string"
          },
          "dispatch_claim_count": {
            "minimum": 0.0,
            "title": "Dispatch Claim Count",
            "type": "integer"
          },
          "empty_outbox": {
            "title": "Empty Outbox",
            "type": "boolean"
          },
          "payload_contract": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Payload Contract",
            "type": "string"
          },
          "payload_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Payload Sha256",
            "type": "string"
          },
          "pointer_published": {
            "title": "Pointer Published",
            "type": "boolean"
          },
          "pointer_status": {
            "enum": [
              "CURRENT",
              "MISSING",
              "STALE_REPAIRABLE"
            ],
            "title": "Pointer Status",
            "type": "string"
          },
          "runner_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Runner Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/MonitoringDeliveryStateV1"
          },
          "transport": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Transport",
            "type": "string"
          },
          "unresolved_claim_numbers": {
            "items": {
              "type": "integer"
            },
            "maxItems": 64,
            "title": "Unresolved Claim Numbers",
            "type": "array"
          },
          "unresolved_dispatch_claim": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MonitoringOperationsUnresolvedClaimV1"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "delivery_id",
          "runner_id",
          "activation_id",
          "cycle_id",
          "alert_batch_id",
          "destination_id",
          "transport",
          "payload_contract",
          "payload_sha256",
          "empty_outbox",
          "created_at",
          "state",
          "pointer_published",
          "pointer_status",
          "dispatch_claim_count"
        ],
        "title": "MonitoringOperationsDeliveryV1",
        "type": "object"
      },
...
      "MonitoringOperationsProjectionV1": {
        "additionalProperties": false,
        "description": "Versioned, secret-free read model over the configured monitoring chain.",
        "properties": {
          "activation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MonitoringOperationsActivationV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "contract": {
            "const": "monitoring_operations_projection_v1",
            "default": "monitoring_operations_projection_v1",
            "title": "Contract",
            "type": "string"
          },
          "cycle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MonitoringOperationsCycleV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "deliveries": {
            "items": {
              "$ref": "#/components/schemas/MonitoringOperationsDeliveryV1"
            },
            "maxItems": 64,
            "title": "Deliveries",
            "type": "array"
          },
          "deliveries_configured": {
            "title": "Deliveries Configured",
            "type": "boolean"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/MonitoringOperationsJobV1"
            },
            "maxItems": 256,
            "title": "Jobs",
            "type": "array"
          },
          "runner": {
            "$ref": "#/components/schemas/MonitoringOperationsRunnerV1"
          },
          "runner_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Runner Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "watchlist": {
            "$ref": "#/components/schemas/MonitoringStatusV1"
          }
        },
        "required": [
          "runner_id",
          "watchlist",
          "runner",
          "deliveries_configured"
        ],
        "title": "MonitoringOperationsProjectionV1",
        "type": "object"
      },
...
      "ResearchSurfaceSnapshotV1": {
        "additionalProperties": false,
        "description": "Versioned deterministic read-only projection for agents and APIs.",
        "properties": {
          "accepted_adjustment_ids": {
            "items": {
              "type": "string"
            },
            "title": "Accepted Adjustment Ids",
            "type": "array"
          },
          "analysis": {
            "$ref": "#/components/schemas/SurfaceAnalysisView"
          },
          "analysis_id": {
            "minLength": 1,
            "title": "Analysis Id",
            "type": "string"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          },
          "content_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Sha256",
            "type": "string"
          },
          "contract": {
            "const": "research_surface_snapshot_v1",
            "default": "research_surface_snapshot_v1",
            "title": "Contract",
            "type": "string"
          },
          "decision_trace_reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SurfaceArtifactReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "historical_status": {
            "$ref": "#/components/schemas/SurfaceHistoricalStatus"
          },
          "profile_id": {
            "minLength": 1,
            "title": "Profile Id",
            "type": "string"
          },
          "research_report_reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SurfaceArtifactReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source_artifacts": {
            "items": {
              "$ref": "#/components/schemas/SurfaceArtifactReference"
            },
            "minItems": 1,
            "title": "Source Artifacts",
            "type": "array"
          },
          "surface_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Surface Id",
            "type": "string"
          }
        },
        "required": [
          "surface_id",
          "content_sha256",
          "analysis_id",
          "as_of",
          "profile_id",
          "company",
          "analysis",
          "source_artifacts",
          "historical_status"
        ],
        "title": "ResearchSurfaceSnapshotV1",
        "type": "object"
      },
...
      "SurfaceErrorDetail": {
        "additionalProperties": false,
        "description": "Stable machine-readable error detail.",
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "SurfaceErrorDetail",
        "type": "object"
      },
      "SurfaceErrorResponse": {
        "additionalProperties": false,
        "description": "Stable error envelope used by the API adapter.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/SurfaceErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "SurfaceErrorResponse",
        "type": "object"
      },
...
      "SurfaceListResponse": {
        "additionalProperties": false,
        "description": "Deterministically ordered surface metadata response.",
        "properties": {
          "surfaces": {
            "items": {
              "$ref": "#/components/schemas/SurfaceMetadata"
            },
            "title": "Surfaces",
            "type": "array"
          }
        },
        "required": [
          "surfaces"
        ],
        "title": "SurfaceListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only access to explicitly supplied ResearchSurfaceSnapshotV1 artifacts.",
    "title": "Turtle Value Engine Research Surface API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/monitoring/operations": {
      "get": {
        "operationId": "getMonitoringOperations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringOperationsProjectionV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurfaceErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurfaceErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurfaceErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurfaceErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Get Monitoring Operations",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/v1/surfaces": {
      "get": {
        "operationId": "listSurfaces",
        "parameters": [
          {
            "in": "query",
            "name": "primary_listing",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Primary Listing"
            }
          },
          {
            "in": "query",
            "name": "profile_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurfaceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurfaceErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Surfaces",
        "tags": [
          "surfaces"
        ]
      }
    },
    "/v1/surfaces/{surface_id}": {
      "get": {
        "operationId": "getSurface",
        "parameters": [
          {
            "in": "path",
            "name": "surface_id",
            "required": true,
            "schema": {
              "title": "Surface Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchSurfaceSnapshotV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurfaceErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurfaceErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Surface",
        "tags": [
          "surfaces"
        ]
      }
    }
  }
}